Add SNP calling workflow stub
